Made tokens use pointers.

+++ myMath.cpp
@@ -1,6 +1,6 @@
 #include "myMath.h"
 
-int64_t powll(int64_t Base, int Exp)
+int64_t powll(int64_t Base, unsigned long long Exp)
 {
     int64_t count = Base;
     for (int i = 1; i < Exp; i++)
