+++ myMath.h
@@ -3,6 +3,6 @@
 #ifndef CALCULATOR__MYMATH_H
 #define CALCULATOR__MYMATH_H
 
-int64_t powll(int64_t Base, int Exp);
+int64_t powll(int64_t Base, unsigned long long  Exp);
 
 #endif //CALCULATOR__MYMATH_H
